Adding the methods that routes will point to for device views
Once I complete the seperation of JSON parsing and Beer Style
information, these are the keg and fermenter view methods that
routes file will point to in order to obtain their proper
information.

Admittedly this is pure vanity on my part, I could simply use
JSONObject.toJSONString() for this output, but I like to output
it in the "proper" order.