added static for holding the tag names and refactored to use
Building out the structure to make this a little less
thrown-together seeming.  using a static/finals to start denoting
the particular tags for the JSON syntax and growing a more
complete BeerStyles class from there.

Still have not flipped the switch to remove the beer styles
logic from JsonLibrary, because I am a tool that way :-p